+++ addons/formguide/Formguide.php
@@ -16,6 +16,7 @@ namespace addons\formguide;
 
 use think\Addons;
 use think\Db;
+use think\facade\Config;
 
 class Formguide extends Addons
 {
@@ -95,13 +96,18 @@ class Formguide extends Addons
     //卸载
     public function uninstall()
     {
-        //删除模型中建的表
-        $table_list = Db::name('model')->where('module', 'formguide')->field('tablename,id')->select();
-        if ($table_list) {
-            foreach ($table_list as $val) {
-                $tablename = config('database.prefix') . $val['tablename'];
-                Db::execute("DROP TABLE IF EXISTS `{$tablename}`;");
-                Db::name('model_field')->where('modelid', $val['id'])->delete();
+        $droptables = request()->param("droptables");
+        $auth       = \app\admin\service\User::instance();
+        //只有开启调试且为超级管理员才允许删除相关数据库
+        if ($droptables && Config::get("app_debug") && $auth->isAdministrator()) {
+            //删除模型中建的表
+            $table_list = Db::name('model')->where('module', 'formguide')->field('tablename,id')->select();
+            if ($table_list) {
+                foreach ($table_list as $val) {
+                    $tablename = Config::get('database.prefix') . $val['tablename'];
+                    Db::execute("DROP TABLE IF EXISTS `{$tablename}`;");
+                    Db::name('model_field')->where('modelid', $val['id'])->delete();
+                }
             }
         }
         //删除模型中的表
